refactoring crud operation and added validation logic

+++ server/controllers/empController.js
@@ -4,16 +4,23 @@ export const addEmployee = async (req, res) => {
   try {
     let { name, email, contact, department } = req.body;
     if (!name || !email || !contact || !department) {
-      return res.status(400)
+      res.status(400);
       throw new Error("All Fields are mandatory");
     }
 
     name = name.trim().toLowerCase();
     email = email.trim().toLowerCase();
+    department = department.trim();
+
+    if(contact.length !== 10){
+      res.status(400);
+      throw new Error("contact contain 10 numbers!!")
+    }
 
     const existingEmp = await empModel.findOne({ email });
     if (existingEmp) {
-      return res.status(400).json({ success: false, message: "Email already exists" });
+      res.status(400)
+      throw new Error("Email already exists")
     }
 
     const emp = new empModel({
@@ -29,7 +36,9 @@ export const addEmployee = async (req, res) => {
     });
   } catch (err) {
     console.log("add employee error:", err);
-    return res.json({ success: false, message: "error occurred!" });
+    // return res.json({ success: false, message: "error occurred!" });
+    res.status(400);
+    throw new Error(err.message)
   }
 };
 
@@ -38,31 +47,44 @@ export const getEmployees = async (req, res) => {
     const emps = await empModel.find();
     return res.json({ success: true, employees: emps });
   } catch (err) {
-    console.log("add employee error:", err);
-    return res.json({ success: false, message: "error occurred!" });
+    console.log("get employees error:", err);
+    res.status(400);
+    throw new Error(err.message)
+
   }
 };
 
 export const getEmployeeById = async (req, res) => {
   try {
     const id = req.params.id;
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        res.status(400);
+        throw new Error("invalid id!!")
+    }
     const emp = await empModel.findById(id);
     if (!emp) {
-      return res.json({ success: false, message: "employee not found !!" });
+      res.status(404);
+      throw new Error("employee not found !!")
     }
     return res.json({ success: true, employee: emp });
   } catch (err) {
-    console.log("add employee error:", err);
-    return res.json({ success: false, message: "error occurred!" });
+    console.log("get employee error:", err);
+   res.status(400);
+    throw new Error(err.message)
   }
 };
 
 export const updateEmployee = async (req, res) => {
   try {
     const id = req.params.id;
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        res.status(400);
+        throw new Error("invalid id!!")
+    }
     const emp = await empModel.findById(id);
     if (!emp) {
-      return res.json({ success: false, message: "employee not found !!" });
+       res.status(404);
+      throw new Error("employee not found !!")
     }
     const updatedEmployee = await empModel.findByIdAndUpdate(id, req.body, {
       new: true,
@@ -70,22 +92,29 @@ export const updateEmployee = async (req, res) => {
 
     return res.json({ success: true, employee: updatedEmployee });
   } catch (err) {
-    console.log("add employee error:", err);
-    return res.json({ success: false, message: "error occurred!" });
+    console.log("update employee error:", err);
+   res.status(400);
+    throw new Error(err.message)
   }
 };
 
 export const deleteEmployee = async (req, res) => {
   try {
     const id = req.params.id;
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        res.status(400);
+        throw new Error("invalid id!!")
+    }
     const emp = await empModel.findById(id);
     if (!emp) {
-      return res.json({ success: false, message: "employee not found !!" });
+      res.status(404);
+      throw new Error("employee not found !!")
     }
     const deleted = await empModel.findByIdAndDelete(id);
     return res.json({ success: true, employee: deleted });
   } catch (err) {
-    console.log("add employee error:", err);
-    return res.json({ success: false, message: "error occurred!" });
+    console.log("delete employee error:", err);
+   res.status(400);
+    throw new Error(err.message)
   }
 };
